split of file
split the file into 2. One is llm explanation (the model identifies and explains the mistakes in the summaries), and the other is model evaluation, which is used to evaluate the output against the label. There is a third file, for llm as a judge for evaluation, the file model evaluation will be used for all other metrics